assistant/pyproject.toml
[project]
name = "assistant"
version = "0.1.0"
description = ""
authors = [
    {name = "Siddhanta Acharya",email = "siddhu12980@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "selenium (>=4.29.0,<5.0.0)",
    "SpeechRecognition>=3.10.0",
    "PyAudio>=0.2.14",
    "pyttsx3>=2.90",
    "openai>=1.12.0",
    "python-dotenv>=1.0.0"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
